+++ src/main/java/server/business/service/SchoolClassService.java
@@ -4,9 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import server.data.entity.SchoolClass;
 import server.data.repository.SchoolClassRepository;
+import server.utils.exception.internalerror.DatabaseOperationExceptionCustom;
+import server.utils.exception.notfound.SchoolClassCustomNotFoundException;
 
 import java.sql.SQLException;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -17,18 +18,31 @@ public class SchoolClassService {
 
     public SchoolClass createClass(SchoolClass schoolClass) {
         schoolClass.setId(UUID.randomUUID());
-        return schoolClassRepository.save(schoolClass);
+        try {
+            return schoolClassRepository.save(schoolClass);
+        } catch (SQLException e) {
+            throw new DatabaseOperationExceptionCustom("Failed to create school class");
+        }
     }
 
     public void deleteClass(UUID id) {
-        schoolClassRepository.deleteClass(id);
+        try {
+            if (schoolClassRepository.doesSchoolClassExist(id)) {
+                schoolClassRepository.deleteClass(id);
+            } else {
+                throw new SchoolClassCustomNotFoundException("Class not found with ID: " + id);
+            }
+        } catch (SQLException e) {
+            throw new DatabaseOperationExceptionCustom("Failed to delete school class with ID: " + id);
+        }
     }
 
-    public Optional<SchoolClass> findClassById(UUID id) {
-        return schoolClassRepository.findClassById(id);
-    }
-
-    public boolean doesClassExist(UUID id) {
-        return schoolClassRepository.doesSchoolClassExist(id);
+    public SchoolClass findClassById(UUID id) {
+        try {
+            return schoolClassRepository.findClassById(id)
+                    .orElseThrow(() -> new SchoolClassCustomNotFoundException("Class not found with ID: " + id));
+        } catch (SQLException e) {
+            throw new DatabaseOperationExceptionCustom("Failed to find school class with ID: " + id);
+        }
     }
 }
